[cleanup] move our boost/sha1 copy to websocket
