+++ xbmc/network/websocket/sha1.hpp
